Reorganise the helpers code for endtoend tests
